restructured controller classes and deleted duplicated hello application
