--- src/main/java/com/example/samplesalad/HelloController.java
@@ -1,104 +1,0 @@
-package com.example.samplesalad;
-
-import javafx.animation.FadeTransition;
-import javafx.animation.TranslateTransition;
-import javafx.fxml.FXML;
-import javafx.fxml.FXMLLoader;
-import javafx.fxml.Initializable;
-import javafx.scene.Parent;
-import javafx.scene.image.ImageView;
-import javafx.scene.input.MouseEvent;
-import javafx.scene.layout.AnchorPane;
-import javafx.scene.layout.BorderPane;
-import javafx.util.Duration;
-import java.util.logging.Logger;
-import java.net.URL;
-import java.util.ResourceBundle;
-
-public class HelloController implements IController {
-    @FXML
-    protected ImageView exit, menu;
-
-    @FXML
-    protected AnchorPane pane1, pane2;
-
-    @FXML
-    private BorderPane bp;
-
-    @FXML
-    protected AnchorPane contentPane;
-
-    public void initialize(URL url, ResourceBundle resourceBundle) {
-        exit.setOnMouseClicked(event -> {
-            System.exit(0);
-        });
-
-        pane1.setVisible(false);
-        pane1.setMouseTransparent(true); // Initially make pane1 non-interactive
-
-        FadeTransition fadeTransition = new FadeTransition(Duration.seconds(0.5), pane1);
-        fadeTransition.setFromValue(1);
-        fadeTransition.setToValue(0);
-        fadeTransition.play();
-
-        TranslateTransition translateTransition = new TranslateTransition(Duration.seconds(0.5), pane2);
-        translateTransition.setByX(-600);
-        translateTransition.play();
-
-        menu.setOnMouseClicked(event -> {
-            pane1.setVisible(true);
-            pane1.setMouseTransparent(false); // Make pane1 interactive
-            FadeTransition fadeTransition1 = new FadeTransition(Duration.seconds(0.5), pane1);
-            fadeTransition1.setFromValue(0);
-            fadeTransition1.setToValue(1);
-            fadeTransition1.play();
-
-            TranslateTransition translateTransition1 = new TranslateTransition(Duration.seconds(0.5), pane2);
-            translateTransition1.setByX(+600);
-            translateTransition1.play();
-        });
-
-        pane1.setOnMouseClicked(event -> {
-            FadeTransition fadeTransition1 = new FadeTransition(Duration.seconds(0.5), pane1);
-            fadeTransition1.setFromValue(1);
-            fadeTransition1.setToValue(0);
-            fadeTransition1.play();
-
-            fadeTransition1.setOnFinished(event1 -> {
-                pane1.setVisible(false);
-                pane1.setMouseTransparent(true); // Make pane1 non-interactive
-            });
-
-            TranslateTransition translateTransition1 = new TranslateTransition(Duration.seconds(0.5), pane2);
-            translateTransition1.setByX(-600);
-            translateTransition1.play();
-        });
-    }
-
-    @FXML
-    private void login(MouseEvent event) {
-        loadPage("login");
-    }
-
-    @FXML
-    private void goToMain(MouseEvent event) {
-        loadPage("hello-view");
-    }
-
-    public void loadPage(String page) {
-        try {
-            FXMLLoader loader = new FXMLLoader(getClass().getResource(page + ".fxml"));
-            Parent root = loader.load();
-            if (page.equals("hello-view")) {
-                AnchorPane newContentPane = (AnchorPane) root.lookup("#contentPane");
-                if (newContentPane != null) {
-                    contentPane.getChildren().setAll(newContentPane.getChildren());
-                }
-            } else {
-                contentPane.getChildren().setAll(root);
-            }
-        } catch (java.io.IOException ex) {
-            Logger.getLogger(HelloController.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        }
-    }
-}
