+++ src/main/java/com/example/samplesalad/controller/IController.java
@@ -1,0 +1,11 @@
+package com.example.samplesalad.controller;
+
+import javafx.fxml.Initializable;
+
+public abstract interface IController extends Initializable {
+
+    @Override
+    public void initialize(java.net.URL url, java.util.ResourceBundle resourceBundle);
+
+    abstract void loadPage(String page);
+}
