+++ src/main/resources/com/example/samplesalad/library-condensed.fxml
@@ -10,7 +10,7 @@
 <?import javafx.scene.layout.AnchorPane?>
 <?import javafx.scene.layout.StackPane?>
 
-<SplitPane fx:controller = "com.example.samplesalad.libraryController" dividerPositions="0.36454849498327757" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/21" xmlns:fx="http://javafx.com/fxml/1">
+<SplitPane fx:controller = "com.example.samplesalad.controller.libraryController" dividerPositions="0.36454849498327757" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/21" xmlns:fx="http://javafx.com/fxml/1">
         <StackPane>
             <AnchorPane minHeight="0.0" minWidth="0.0" prefHeight="398.0" prefWidth="242.0">
                     <Button id="expandButton" layoutX="187.0" layoutY="-1.0" mnemonicParsing="false" onMouseClicked="#onSideBarClicked" prefHeight="400.0" prefWidth="15.0" text="&gt;" />
