+++ src/main/resources/com/example/samplesalad/library-expanded.fxml
@@ -12,7 +12,7 @@
 
 <SplitPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="400.0"
            prefWidth="600.0" xmlns="http://javafx.com/javafx/21" xmlns:fx="http://javafx.com/fxml/1"
-           fx:controller="com.example.samplesalad.libraryController">
+           fx:controller="com.example.samplesalad.controller.libraryController">
         <AnchorPane minHeight="0.0" minWidth="0.0" prefHeight="398.0" prefWidth="242.0">
             <Button id="expandButton" layoutX="568.0" layoutY="-1.0" mnemonicParsing="false"
                     onMouseClicked="#onSideBarClicked" prefHeight="400.0" prefWidth="31.0" text="&lt;"/>
